+++ tests/test_SpectralData.cpp
@@ -17,7 +17,7 @@
 void init_Spectrum( rta::core::Spectrum &spectrum )
 {
     for ( size_t i = 0; i < spectrum.values.size(); i++ )
-        spectrum.values[i] = i;
+        spectrum.values[i] = static_cast<double>( i );
 }
 
 void check_Spectrum(
